requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysqlclient
PyJWT==2.10.1
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
gunicorn
psycopg==3.1.18
# Variáveis de ambiente
python-dotenv==1.0.0

# Segurança
cryptography==41.0.7
Flask-Talisman==1.1.0
Flask-Limiter==3.5.0
bleach==6.1.0
email-validator==2.1.0